remove jcr-specific APIs from AbstractPostOperation
By promoting impl/operations.AbstractPostOperation to the exported package
